Switch from f64 to f32
Allows the activity to work without casting

stylistic choice - could also change the activity to take f64.

Unknown which is better.

+++ examples/custom_types/structs/structs.rs
@@ -2,12 +2,12 @@
 struct Nil;
 
 // A tuple struct
-struct Pair(i32, f64);
+struct Pair(i32, f32);
 
 // A struct with two fields
 struct Point {
-    x: f64,
-    y: f64,
+    x: f32,
+    y: f32,
 }
 
 // Structs can be reused as fields of another struct
